Add Migrations

+++ Domain/Entities/OrderOption.cs
@@ -8,15 +8,14 @@ namespace Domain.Entities
 {
     public class OrderOption
     {
-
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double Price { get; set; }
-        public int Discount { get; set; }
+        public decimal Price { get; set; }
+        public decimal Discount { get; set; }
         public DateOnly DiscountStart { get; set; }
         public DateOnly DiscountEnd { get; set; }
         public int ReportNumber { get; set; }
-        public int TotalAmount { get; set; }
+        public decimal TotalAmount { get; set; }
     }
 }
